requirements.txt
anyio==3.7.1
certifi==2023.7.22
dnspython==2.4.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pymongo==4.4.1
python-dotenv==1.0.0
python-telegram-bot==20.4
sniffio==1.3.0
